This is very frustrating and I think an issue with vcpkg rather than the ngl build. Under Mac we have an issue that we need to add Both IMath and OpenEXR packages when using OpenImageIO.
At present this is not needed under windows or linux (Centos tested so far). but I think that this may change as I have seen it being needed then not depending upon the version of vcpkg.

I think as of this commit I have tested it with the latest 2020.06.15 version of vcpkg on mac, windows and linux and the mac version is playing up. We also need to modify the libsquish path as well. I will submit an issue on the vcpkg page and see what happens.